move lifecycle mixin under supervisor
